Update to 15.0.3
This commit includes a patch to the QEMU test suite.
The tcg/multiarch/linux/linux-test.c file fails to
compile under -Wall with LLVM 15.0.3

This is likely due to [recent improvements][1] to the
-Wno-unused-but-set-variable diagnostic to fix false
negatives when post/pre increment/decrement operator
are involved.

The patch simply deletes the unused variable since
the semantics of the test will remain the same.

[1]: https://github.com/llvm/llvm-project/issues/54595#issuecomment-1081022049

Signed-off-by: Mukilan Thiyagarajan <quic_mthiyaga@quicinc.com>