Change to use cmake cache files
cmake cache files are a popular idiom with cmake projects in general
and llvm-project specifically.  If we have luck using these in
the toolchain_for_hexagon repo, perhaps we could contribute them
to llvm-project to make it easier for others to reproduce this build
and easier for other llvm developers to see how our toolchain is built.

Signed-off-by: Brian Cain <bcain@quicinc.com>