Update to Ubuntu 20.04
We have been living on borrowed time -- QEMU has for some time
required Ubuntu 20.04 to build and recently has made changes that prevent
QEMU from building on the glibc from Ubuntu 18.04.   The ubuntu-based
toolchain built by this repo will depend on 20.04.

Update to 20.04 and make corresponding changes to the docker config
to support differences between 18.04 and 20.04.

Signed-off-by: Brian Cain <bcain@quicinc.com>