Add default clang symlinks to CLANG_LINKS_TO_CREATE
Since this cache value overrides the defaults, we end up with `clang` linked
to `clang-20`, and some `${triple}-clang*` links, but we're missing `clang++`.
This makes for a toolchain with inconsistent behavior when used in someone's
`$PATH`.

We'll add the default symlinks to our list so that C and C++ programs
are both built as expected when `clang` and `clang++` are invoked.

Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>