[cmake] Override runtimes' feature detection
In https://github.com/llvm/llvm-project/commit/6c4267fb1779bc5550bb413f33250f9365acfbc6,
there's a `target_compile_features()` directive that appears to cause the
runtimes build to check the toolchain for C++17 support.  This appears
to depend on the C++ library being present, which is not possible because
we haven't built it yet.

Either this is a bug in the runtimes' build mechanism or it's the intended
behavior and we need to special-case our bootstrap build.  Regardless,
this change makes sense for the time being.

See also:
* https://github.com/llvm/llvm-project/issues/114591
* https://github.com/llvm/llvm-project/pull/123184
* https://discourse.llvm.org/t/bootstrapping-cross-compiler-runtimes-build-failure/84194

Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>