Build toolchain with PIC
This will be required for eld.

Some of the cross-built toolchains fail, because a limitation with zig's
use of libclang.  They fail with `error: unsupported linker arg: -Bsymbolic-functions`.
Bypassing that failure by disabling the clang-shlib results in a link
failure (see below).  So for now we'll just use non-PIC for the
toolchain targets that fail.

```
LLD Link... ld.lld: error: attempted static link of dynamic object lib/libclang.so.20.1.0-rc3
```

Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>