Fix compiler specification for musl build
Instead of using `${CROSS_COMPILE}${CC}` for the compiler as I expected, the
musl build seems to just use ${CC}.

When switching the build to v68 to accomodate eld's oldest supported
arch revision, I found that the build failed because v68 wasn't
supported by the compiler.  This was due to the fact that it was using
clang -- meaning the host clang compiler and not the just-built clang.

Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>