Add some feedback before cloning
get-src-repos.sh was written to be executed in a noninteractive build
environment.  And since the clones are all "--quiet", there's no
feedback to an interactive user when running the program.  Since it
appears to just hang for a long time, let's make it clear that we've
started out the repo clone operation.

Signed-off-by: Brian Cain <brian.cain@oss.qualcomm.com>